requirements.txt
streamlit>=1.20
langchain>=0.1.0
openai>=1.0.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pgvector==0.4.1
gTTS>=2.3.0
python-telegram-bot>=20.0
tiktoken>=0.4.0
whisper-openai==1.0.0
pytest>=7.0

# In-browser recording
streamlit-webrtc>=0.54.0
numpy>=1.24
soundfile>=0.12.1
av>=10.0.0
streamlit>=1.20
streamlit-chat>=0.0.4